Moved Scope into its own file
