+++ sql/analyzer/analyzer.go
@@ -167,102 +167,6 @@ type Analyzer struct {
 	Catalog *sql.Catalog
 }
 
-// Scope of the analysis being performed, used when analyzing subqueries to give such analysis access to outer scope.
-type Scope struct {
-	// Stack of nested node scopes, with innermost scope first
-	nodes []sql.Node
-}
-
-// newScope creates a new Scope object with the additional innermost Node context. When constructing with a subquery,
-// the Node given should be the sibling Node of the subquery.
-func (s *Scope) newScope(node sql.Node) *Scope {
-	if s == nil {
-		return &Scope{[]sql.Node{node}}
-	}
-	newNodes := make([]sql.Node, len(s.nodes)+1)
-	newNodes = append(newNodes, node)
-	newNodes = append(newNodes, s.nodes...)
-	return &Scope{newNodes}
-}
-
-// InnerToOuter returns the scope Nodes in order of innermost scope to outermost scope.
-func (s *Scope) InnerToOuter() []sql.Node {
-	if s == nil {
-		return nil
-	}
-	return s.nodes
-}
-
-// OuterToInner returns the scope nodes in order of outermost scope to innermost scope.
-func (s *Scope) OuterToInner() []sql.Node {
-	if s == nil {
-		return nil
-	}
-	reversed := make([]sql.Node, len(s.nodes))
-	for i := range s.nodes {
-		reversed[i] = s.nodes[len(s.nodes)-i-1]
-	}
-	return reversed
-}
-
-// SchemaLength returns the length of the scope part of the schema for nodes subject to this scope. Rows for this node
-// will have all scoped schemas pre-pended to them, so that the results of the node under analysis begins with index
-// SchemaLength().
-func (s *Scope) SchemaLength() int {
-	if s == nil {
-		return 0
-	}
-
-	// Nodes in the scope might not be resolved, so we can't call Schema() on them. Instead, examine them manually.
-	var length int
-	for _, n := range s.OuterToInner() {
-		switch n := n.(type) {
-		case *plan.Project:
-			length += len(n.Projections)
-		default:
-			// TODO: log this
-			// panic(fmt.Sprintf("Unsupported scope node %T", n))
-		}
-	}
-	return length
-}
-
-// Schema returns the equivalent schema of this scope, which consists of the schemas of all constituent scope nodes
-// concatenated from outer to inner. Because we can only calculate the Schema() of nodes that are Resolved(), this
-// method fills in place holder columns as necessary.
-func (s *Scope) Schema() sql.Schema {
-	var schema sql.Schema
-	for _, n := range s.OuterToInner() {
-		if n.Resolved() {
-			schema = append(schema, n.Schema()...)
-			continue
-		}
-
-		// If this scope node isn't resolved, we can't use Schema() on it. Instead, assemble an equivalent Schema, with
-		// placeholder columns where necessary, for the purpose of analysis.
-		switch n := n.(type) {
-		case *plan.Project:
-			for _, expr := range n.Projections {
-				var col *sql.Column
-				if expr.Resolved() {
-					col = expression.ExpressionToColumn(expr)
-				} else {
-					// TODO: a new type here?
-					col = &sql.Column{
-						Name:   "",
-						Source: "",
-					}
-				}
-				schema = append(schema, col)
-			}
-		default:
-			// TODO: log this
-			// panic(fmt.Sprintf("Unsupported scope node %T", n))
-		}
-	}
-	return schema
-}
-
 // NewDefault creates a default Analyzer instance with all default Rules and configuration.
 // To add custom rules, the easiest way is use the Builder.
 func NewDefault(c *sql.Catalog) *Analyzer {
